refactored ulog-dropdown - extracted as common component
